requirements.txt
boto3==1.12.39
botocore==1.15.39
cachetools==4.1.0
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.1
docutils==0.15.2
Flask==1.1.2
google-api-core==1.16.0
google-auth==1.13.1
google-cloud-core==1.3.0
google-cloud-speech==1.3.2
google-cloud-texttospeech==1.0.1
google-cloud-translate==2.0.0
googleapis-common-protos==1.51.0
grpcio==1.28.1
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.1
jmespath==0.9.5
MarkupSafe==1.1.1
protobuf==3.11.3
python-dotenv==0.12.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
python-dateutil==2.8.1
pytz==2019.3
requests==2.23.0
rsa==4.0
s3transfer==0.3.3
six==1.14.0
urllib3==1.25.8
Werkzeug==1.0.1

